+++ lib/ts/claims/primitiveArrayClaim.ts
@@ -154,6 +154,52 @@ export class PrimitiveArrayClaim<ValueType> {
                 }
             };
         },
+        includesAny: (
+            val: ValueType[],
+            maxAgeInSeconds: number | undefined = this.defaultMaxAgeInSeconds,
+            id?: string
+        ): SessionClaimValidator => {
+            return {
+                id: id !== undefined ? id : this.id,
+                refresh: ctx => this.refresh(ctx),
+                shouldRefresh: (payload, ctx) =>
+                    this.getValueFromPayload(payload, ctx) === undefined ||
+                    // We know payload[this.id] is defined since the value is not undefined in this branch
+                    (maxAgeInSeconds !== undefined && payload[this.id].t < Date.now() - maxAgeInSeconds * 1000),
+                validate: async (payload, ctx) => {
+                    const claimVal = this.getValueFromPayload(payload, ctx);
+                    if (claimVal === undefined) {
+                        return {
+                            isValid: false,
+                            reason: { message: "value does not exist", expectedToInclude: val, actualValue: claimVal }
+                        };
+                    }
+                    const ageInSeconds = (Date.now() - this.getLastFetchedTime(payload, ctx)!) / 1000;
+                    if (maxAgeInSeconds !== undefined && ageInSeconds > maxAgeInSeconds) {
+                        return {
+                            isValid: false,
+                            reason: {
+                                message: "expired",
+                                ageInSeconds,
+                                maxAgeInSeconds
+                            }
+                        };
+                    }
+                    const claimSet = new Set(claimVal);
+                    const isValid = val.some(v => claimSet.has(v));
+                    return isValid
+                        ? { isValid }
+                        : {
+                              isValid,
+                              reason: {
+                                  message: "wrong value",
+                                  expectedToIncludeAtLeastOneOf: val,
+                                  actualValue: claimVal
+                              }
+                          };
+                }
+            };
+        },
         excludesAll: (
             val: ValueType[],
             maxAgeInSeconds: number | undefined = this.defaultMaxAgeInSeconds,