+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [unreleased]
+- Added `includesAny` validator to Primitive Array claims.
+
+### Breaking change
+- Added `getHost` function to the `WindowHandlerInterface`.
 
 ## [16.0.0] - 2023-02-01